move default options to styles and add separate default options for each module
