+++ cleopatra/array.py
@@ -1,97 +1,32 @@
 """plotting Array."""
-from collections import OrderedDict
 from typing import Any, Union, List
 
 import matplotlib.colors as colors
 import matplotlib.pyplot as plt
 import numpy as np
-
 from numpy_utils.filter import get_indices2
 
 # from matplotlib import gridspec
 from matplotlib.animation import FuncAnimation
 from matplotlib.ticker import LogFormatter
-from cleopatra.styles import Scale, MidpointNormalize
+from cleopatra.styles import DEFAULT_OPTIONS as style_defaults
+from cleopatra.styles import Styles, Scale, MidpointNormalize
 
 DEFAULT_OPTIONS = dict(
-    figsize=(8, 8),
     vmin=None,
     vmax=None,
-    title=None,
-    title_size=15,
-    cbar_length=0.75,
-    orientation="vertical",
-    cmap="coolwarm_r",
-    cbar_label_size=12,
-    cbar_label="Color bar label",
-    rotation=-90,
-    ticks_spacing=5,
     num_size=8,
-    color_scale=1,
-    gamma=0.5,
-    line_scale=0.001,
-    line_threshold=0.0001,
-    bounds=None,
-    midpoint=0,
     display_cell_value=False,
     background_color_threshold=None,
     id_color="green",
     id_size=20,
-    tick_spacing=10,
     precission=2,
 )
+DEFAULT_OPTIONS = style_defaults | DEFAULT_OPTIONS
 
 
 class Array:
     """Map."""
-
-    figure_default_options = dict(
-        ylabel="",
-        xlabel="",
-        legend="",
-        legend_size=10,
-        figsize=(10, 8),
-        labelsize=10,
-        fontsize=10,
-        name="hist.tif",
-        color1="#3D59AB",
-        color2="#DC143C",
-        linewidth=3,
-        Axisfontsize=15,
-    )
-
-    line_styles = OrderedDict(
-        [
-            ("solid", (0, ())),  # 0
-            ("loosely dotted", (0, (1, 10))),  # 1
-            ("dotted", (0, (1, 5))),  # 2
-            ("densely dotted", (0, (1, 1))),  # 3
-            ("loosely dashed", (0, (5, 10))),  # 4
-            ("dashed", (0, (5, 5))),  # 5
-            ("densely dashed", (0, (5, 1))),  # 6
-            ("loosely dashdotted", (0, (3, 10, 1, 10))),  # 7
-            ("dashdotted", (0, (3, 5, 1, 5))),  # 8
-            ("densely dashdotted", (0, (3, 1, 1, 1))),  # 9
-            ("loosely dashdotdotted", (0, (3, 10, 1, 10, 1, 10))),  # 10
-            ("dashdotdotted", (0, (3, 5, 1, 5, 1, 5))),  # 11
-            ("densely dashdotdotted", (0, (3, 1, 1, 1, 1, 1))),  # 12
-            ("densely dashdotdottededited", (0, (6, 1, 1, 1, 1, 1))),  # 13
-        ]
-    )
-
-    marker_style_list = [
-        "--o",
-        ":D",
-        "-.H",
-        "--x",
-        ":v",
-        "--|",
-        "-+",
-        "-^",
-        "--s",
-        "-.*",
-        "-.h",
-    ]
 
     def __init__(self, array: np.ndarray, exclude_value: List = np.nan):
         """Plot array.
@@ -162,53 +97,6 @@ class Array:
         """Default plot options"""
         return self._default_options
 
-    @staticmethod
-    def get_line_style(style: Union[str, int] = "loosely dotted"):
-        """LineStyle.
-
-        Line styles for plotting
-
-        Parameters
-        ----------
-        style : TYPE, optional
-            DESCRIPTION. The default is 'loosely dotted'.
-
-        Returns
-        -------
-        TYPE
-            DESCRIPTION.
-        """
-        if isinstance(style, str):
-            try:
-                return Array.line_styles[style]
-            except KeyError:
-                msg = (
-                    f" The style name you entered-{style}-does not exist please"
-                    "choose from the available styles"
-                )
-                print(msg)
-                print(list(Array.line_styles))
-        else:
-            return list(Array.line_styles.items())[style][1]
-
-    @staticmethod
-    def get_marker_style(style: int):
-        """Marker styles for plotting.
-
-        Parameters
-        ----------
-        style: [int]
-            DESCRIPTION.
-
-        Returns
-        -------
-        TYPE
-            DESCRIPTION.
-        """
-        if style > len(Array.marker_style_list) - 1:
-            style = style % len(Array.marker_style_list)
-        return Array.marker_style_list[style]
-
     def get_ticks(self) -> np.ndarray:
         """get list of ticks for the color bar"""
         ticks_spacing = self.default_options["ticks_spacing"]
@@ -235,6 +123,8 @@ class Array:
 
         Parameters
         ----------
+        ax: [axes]
+            matplotlib figure axes.
         arr: [array]
             numpy array.
         ticks: [list]
@@ -836,7 +726,7 @@ class Array:
             Y1[:, 1],
             zorder=1,
             color=color1,
-            linestyle=Array.get_line_style(0),
+            linestyle=Styles.get_line_style(0),
             linewidth=linewidth,
             label="Model 1 Output1",
         )
@@ -857,7 +747,7 @@ class Array:
                     Y1_2[:, i],
                     zorder=1,
                     color=color2,
-                    linestyle=Array.get_line_style(i),
+                    linestyle=Styles.get_line_style(i),
                     linewidth=linewidth,
                     label=label[i - 1],
                 )
@@ -867,7 +757,7 @@ class Array:
             Y2[:, 1],
             zorder=1,
             color=color3,
-            linestyle=Array.get_line_style(6),
+            linestyle=Styles.get_line_style(6),
             linewidth=2,
             label="Output1-Diff",
         )
@@ -887,7 +777,7 @@ class Array:
                     Y2_2[:, i],
                     zorder=1,
                     color=color2,
-                    linestyle=Array.get_line_style(i),
+                    linestyle=Styles.get_line_style(i),
                     linewidth=linewidth,
                     label=label[i - 1],
                 )
