+++ cleopatra/styles.py
@@ -1,6 +1,125 @@
 """style related functionality"""
+from collections import OrderedDict
+from typing import Union
 import matplotlib.colors as colors
 import numpy as np
+
+DEFAULT_OPTIONS = dict(
+    figsize=(8, 8),
+    title=None,
+    title_size=15,
+    ylabel="",
+    xlabel="",
+    legend="",
+    legend_size=10,
+    color_1="#3D59AB",
+    color_2="#DC143C",
+    line_width=3,
+    cbar_length=0.75,
+    orientation="vertical",
+    cmap="coolwarm_r",
+    cbar_label_size=12,
+    cbar_label="Color bar label",
+    rotation=-90,
+    ticks_spacing=5,
+    color_scale=1,
+    gamma=0.5,
+    line_scale=0.001,
+    line_threshold=0.0001,
+    bounds=None,
+    midpoint=0,
+)
+
+# figure_default_options = dict(
+#         labelsize=10,
+#         fontsize=10,
+#         name="hist.tif",
+#         Axisfontsize=15,
+#     )
+
+
+class Styles:
+    """Styles"""
+
+    line_styles = OrderedDict(
+        [
+            ("solid", (0, ())),  # 0
+            ("loosely dotted", (0, (1, 10))),  # 1
+            ("dotted", (0, (1, 5))),  # 2
+            ("densely dotted", (0, (1, 1))),  # 3
+            ("loosely dashed", (0, (5, 10))),  # 4
+            ("dashed", (0, (5, 5))),  # 5
+            ("densely dashed", (0, (5, 1))),  # 6
+            ("loosely dashdotted", (0, (3, 10, 1, 10))),  # 7
+            ("dashdotted", (0, (3, 5, 1, 5))),  # 8
+            ("densely dashdotted", (0, (3, 1, 1, 1))),  # 9
+            ("loosely dashdotdotted", (0, (3, 10, 1, 10, 1, 10))),  # 10
+            ("dashdotdotted", (0, (3, 5, 1, 5, 1, 5))),  # 11
+            ("densely dashdotdotted", (0, (3, 1, 1, 1, 1, 1))),  # 12
+            ("densely dashdotdottededited", (0, (6, 1, 1, 1, 1, 1))),  # 13
+        ]
+    )
+
+    marker_style_list = [
+        "--o",
+        ":D",
+        "-.H",
+        "--x",
+        ":v",
+        "--|",
+        "-+",
+        "-^",
+        "--s",
+        "-.*",
+        "-.h",
+    ]
+
+    @staticmethod
+    def get_line_style(style: Union[str, int] = "loosely dotted"):
+        """LineStyle.
+
+        Line styles for plotting
+
+        Parameters
+        ----------
+        style : TYPE, optional
+            DESCRIPTION. The default is 'loosely dotted'.
+
+        Returns
+        -------
+        TYPE
+            DESCRIPTION.
+        """
+        if isinstance(style, str):
+            try:
+                return Styles.line_styles[style]
+            except KeyError:
+                msg = (
+                    f" The style name you entered-{style}-does not exist please"
+                    "choose from the available styles"
+                )
+                print(msg)
+                print(list(Styles.line_styles))
+        else:
+            return list(Styles.line_styles.items())[style][1]
+
+    @staticmethod
+    def get_marker_style(style: int):
+        """Marker styles for plotting.
+
+        Parameters
+        ----------
+        style: [int]
+            DESCRIPTION.
+
+        Returns
+        -------
+        TYPE
+            DESCRIPTION.
+        """
+        if style > len(Styles.marker_style_list) - 1:
+            style = style % len(Styles.marker_style_list)
+        return Styles.marker_style_list[style]
 
 
 class Scale:
